+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - `plot_learning_curve` method of module `weka.plot.classifiers` now accepts a list of test sets;
   `*` is index of test set in label template string
+- added `missing_value()` methods to `weka.core.dataset` module and `Instance` class
 - ...